Changed the definition type and fixed logical bug in SymbolTableGenerator.

+++ src/entities/SymbolTable.java
@@ -5,17 +5,17 @@ import java.util.Set;
 
 public class SymbolTable {
 
-    private final HashMap<String, Integer> symbolTable;
+    private final HashMap<String, Short> symbolTable;
 
     public SymbolTable(){
         this.symbolTable = new HashMap<>();
     }
 
-    public void addSymbol(String label, int definitionPoint){
+    public void addSymbol(String label, short definitionPoint){
         symbolTable.put(label, definitionPoint);
     }
 
-    public int findDefinitionPoint(String label){
+    public short findDefinitionPoint(String label){
         return symbolTable.get(label);
     }
